Providing adapter
