+++ officefloor/cabinet/officecabinet_common/src/main/java/net/officefloor/cabinet/common/CabinetUtil.java
@@ -26,6 +26,8 @@ import java.util.UUID;
 import net.officefloor.cabinet.Document;
 import net.officefloor.cabinet.Document.DocumentNameContext;
 import net.officefloor.cabinet.Document.DocumentNamer;
+import net.officefloor.cabinet.common.key.DocumentKey;
+import net.officefloor.cabinet.common.key.FieldDocumentKey;
 import net.officefloor.cabinet.Key;
 import net.officefloor.cabinet.OfficeCabinet;
