+++ officefloor/cabinet/officecabinet_common/src/main/java/net/officefloor/cabinet/common/key/DocumentKey.java
@@ -1,4 +1,4 @@
-package net.officefloor.cabinet.common;
+package net.officefloor.cabinet.common.key;
 
 import net.officefloor.cabinet.Document;
