+++ officefloor/cabinet/officecabinet_common/src/main/java/net/officefloor/cabinet/common/manage/ManagedDocument.java
@@ -1,4 +1,4 @@
-package net.officefloor.cabinet.common;
+package net.officefloor.cabinet.common.manage;
 
 import net.officefloor.cabinet.Document;
