+++ officefloor/cabinet/officecabinet_common/src/main/java/net/officefloor/cabinet/common/manage/ManagedDocumentState.java
@@ -1,4 +1,4 @@
-package net.officefloor.cabinet.common;
+package net.officefloor.cabinet.common.manage;
 
 /**
  * State of the {@link ManagedDocument}.
