+++ officefloor/cabinet/officecabinet_common/src/main/java/net/officefloor/cabinet/common/metadata/InternalDocument.java
@@ -1,4 +1,4 @@
-package net.officefloor.cabinet.common;
+package net.officefloor.cabinet.common.metadata;
 
 import net.officefloor.cabinet.Document;
 import net.officefloor.cabinet.Key;
