+++ officefloor/cabinet/officecabinet_cosmosdb/src/main/java/net/officefloor/cabinet/cosmosdb/CosmosOfficeCabinetArchive.java
@@ -20,12 +20,14 @@ public class CosmosOfficeCabinetArchive<D> implements OfficeCabinetArchive<D> {
 	/**
 	 * Instantiate.
 	 * 
+	 * @param adapter        {@link CosmosOfficeCabinetAdapter}.
 	 * @param documentType   Document type.
 	 * @param cosmosDatabase {@link CosmosDatabase}.
 	 * @throws Exception If fails to instantiate {@link OfficeCabinet}.
 	 */
-	public CosmosOfficeCabinetArchive(Class<D> documentType, CosmosDatabase cosmosDatabase) throws Exception {
-		this.metaData = new CosmosDocumentMetaData<>(documentType, cosmosDatabase);
+	public CosmosOfficeCabinetArchive(CosmosOfficeCabinetAdapter adapter, Class<D> documentType,
+			CosmosDatabase cosmosDatabase) throws Exception {
+		this.metaData = new CosmosDocumentMetaData<>(adapter, documentType, cosmosDatabase);
 	}
 
 	/*
