+++ officefloor/cabinet/officecabinet_dynamodb/src/main/java/net/officefloor/cabinet/dynamo/DynamoOfficeCabinet.java
@@ -30,7 +30,7 @@ import com.amazonaws.services.dynamodbv2.document.spec.BatchWriteItemSpec;
 
 import net.officefloor.cabinet.OfficeCabinet;
 import net.officefloor.cabinet.common.AbstractOfficeCabinet;
-import net.officefloor.cabinet.common.InternalDocument;
+import net.officefloor.cabinet.common.metadata.InternalDocument;
 import net.officefloor.test.UsesDockerTest;
 
 /**
