+++ officefloor/cabinet/officecabinet_firestore/src/main/java/net/officefloor/cabinet/firestore/FirestoreOfficeCabinet.java
@@ -9,7 +9,7 @@ import com.google.cloud.firestore.Firestore;
 
 import net.officefloor.cabinet.OfficeCabinet;
 import net.officefloor.cabinet.common.AbstractOfficeCabinet;
-import net.officefloor.cabinet.common.InternalDocument;
+import net.officefloor.cabinet.common.metadata.InternalDocument;
 
 /**
  * {@link Firestore} {@link OfficeCabinet}.
